[SP-25] refactor: recommend domain 으로 변경
