+++ src/main/java/com/cupid/jikting/recommend/dto/HobbyResponse.java
@@ -1,4 +1,4 @@
-package com.cupid.jikting.meeting.dto;
+package com.cupid.jikting.recommend.dto;
 
 import lombok.AccessLevel;
 import lombok.AllArgsConstructor;
@@ -10,10 +10,7 @@ import lombok.NoArgsConstructor;
 @Builder
 @AllArgsConstructor(access = AccessLevel.PROTECTED)
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
-public class ImageResponse {
+public class HobbyResponse {
 
-	private boolean isMain;
-	private Long memberId;
-	private String url;
-
+	private String hobby;
 }
